Rename new files.
