+++ src/estimagic/visualization/deviation_plot.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import plotly.express as px
 
-from estimagic.benchmarking.process_benchmark_results_new import (
+from estimagic.benchmarking.process_benchmark_results import (
     process_benchmark_results,
 )
 from estimagic.config import PLOTLY_TEMPLATE
